feat: enhance threedotloading + refactory resultsection + update threedot and implement text split on explainsection
