+++ .vscode/settings.json
@@ -2,9 +2,11 @@
     "cSpell.words": [
         "docco",
         "drawio",
+        "Explaination",
         "Kwargs",
         "langchain",
         "rehype",
+        "SSRVP",
         "syargpt",
         "Tavily",
         "unoptimized",
