+++ frontend/src/components/HistorySection/index.ts
@@ -1,0 +1,3 @@
+import HistorySection from "./HistorySection"
+
+export default HistorySection
